Remove all compiler warnings
Libraries in general should never warn. Even at this early stage, it's
worth just getting these out of the way. The unused fields will of
course be reversed in the future, but it's a minimal change to show that
we know we're not using them at the moment.

Since `io::stdout()::flush()` can fail, but I don't think the reporters
should panic, I've moved that repsonsibility up a level, and changed
the `Reporter` trait to return an `io::Result`.